+++ content/files/cosi-install.sh
@@ -916,8 +916,7 @@ cosi_initialize() {
 
     hn=$(hostname)
     if [[ -z "${hn:-}" && -z "${cosi_host_target:-}" ]]; then
-        echo "cosi requires system hostname to be set or overridden with --target"
-        exit 1
+        fatal "cosi requires system hostname to be set or overridden with --target"
     fi
 
     #
@@ -985,6 +984,8 @@ cosi_register() {
     local cosi_register_cmd="register"
     local cosi_register_opt=""
     local install_reverse="${cosi_dir}/bin/reverse_install.sh"
+    local agent_config="$agent_dir/etc/circonus-agent.yaml"
+    local agent_bin="${agent_dir}/sbin/circonus-agentd"
 
     echo
     __fetch_cosi_tool
@@ -1000,23 +1001,26 @@ cosi_register() {
     [[ -x "$cosi_script" ]] || fail "Unable to find cosi command '${cosi_script}'"
 
     echo
+    log "### Creating base circonus-agent configuration"
+    echo
+    $agent_bin --check-metric-streamtags --check-tags="os:${cosi_os_type},arch:${cosi_os_arch},distro:${cosi_os_dist}-${cosi_os_vers}" --show-config=yaml > $agent_config
+    [[ $? -eq 0 ]] || fail "Error creating base circonus-agent configuration"
+    __restart_agent
+
+    echo
     log "### Running COSI registration ###"
     echo
     log_only "running: $cosi_script" "$cosi_register_cmd"
     "$cosi_script" "$cosi_register_cmd" | tee -a $cosi_install_log
     [[ ${PIPESTATUS[0]} -eq 0 ]] || fail "Errors encountered during registration."
 
-
     if [[ "${cosi_agent_mode}" == "reverse" ]]; then
         echo
         log "### Enabling ${cosi_agent_mode} mode for agent ###"
         echo
-        local agent_bin="${agent_dir}/sbin/circonus-agentd"
-        if [[ -x $agent_bin ]]; then
-            $agent_bin --reverse --check-id="cosi" --api-key="cosi" --api-app="cosi" --check-metric-streamtags --check-tags="os:${cosi_os_type},arch:${cosi_os_arch},distro:${cosi_os_dist}-${cosi_os_vers}" --show-config=yaml > $agent_dir/etc/circonus-agent.yaml
-            [[ $? -eq 0 ]] || fail "Error updating circonus-agent configuration"
-            __restart_agent
-        fi
+        $agent_bin --reverse --check-id="cosi" --api-key="cosi" --api-app="cosi" --show-config=yaml --check-metric-streamtags --check-tags="os:${cosi_os_type},arch:${cosi_os_arch},distro:${cosi_os_dist}-${cosi_os_vers}" > $agent_config
+        [[ $? -eq 0 ]] || fail "Error updating circonus-agent configuration"
+        __restart_agent
     fi
 
     echo
@@ -1049,6 +1053,19 @@ cosi_register() {
 
 cosi_install() {
     cosi_initialize "$@"
+    #
+    # short circuit if NAD installation detected
+    #
+    if [[ -d "${base_dir}/nad" || -f "${base_dir}/sbin/nad.js" ]]; then
+        echo
+        echo
+        log "*** Previous NAD installation detected ***"
+        log "Please use the following crconus-agent installation instructions:"
+        log "https://github.com/circonus-labs/circonus-agent#quick-start"
+        echo
+        echo
+        exit
+    fi
     cosi_verify_os
     cosi_check_agent
     cosi_register